refactor: replace all related to pages components into pages folder
