+++ src/components/containers/ProfileInfo/ProfileInfo.tsx
@@ -4,7 +4,7 @@ import Grid from "@mui/material/Grid";
 import {AvatarProfileInput} from "@containers/AvatarProfileInput/AvatarProfileInput";
 import {ProfileUserInfo} from "@view/ProfileUserInfo/ProfileUserInfo";
 import {ProfileForm} from "@containers/ProfileForm/ProfileForm";
-import {ProgressSpinner} from "@/components/view/ProgressSpinner/ProgressSpinner";
+import {ProgressSpinner} from "@view/ProgressSpinner/ProgressSpinner";
 import {useUserData} from "@/hooks/useUserData";
 
 export const ProfileInfo: FC = () => {
