+++ src/components/view/PageLayout/PageLayout.tsx
@@ -1,8 +1,8 @@
 import {FC, ReactNode} from "react";
 import Container from "@mui/material/Container";
-import {MainHeader} from "@/components/view/MainHeader/MainHeader";
+import {MainHeader} from "@view/MainHeader/MainHeader";
 import {PageLayoutStyle, PageStyle} from "@view/MuiPagesStyles";
-import {BreadCrumbs} from "@/components/view/BreadCrumbs/BreadCrumbs";
+import {BreadCrumbs} from "@view/BreadCrumbs/BreadCrumbs";
 
 interface PageLayoutProps {
   children: ReactNode;
