+++ src/pages/Cvs/Cvs.tsx
@@ -1,5 +1,5 @@
 import {FC} from "react";
-import {TableCvs} from "@/components/containers/TableCvs/TableCvs";
+import {TableCvs} from "@/pages/Cvs/TableCvs/TableCvs";
 import {useCvsQuery} from "@/graphql/hooks/useCvsQuery";
 
 export const Cvs: FC = () => {
