+++ src/pages/Languages/TableLanguages/TableLanguages.tsx
@@ -1,7 +1,7 @@
 import {FC} from "react";
-import {TableHeadTemplate} from "@/components/view/TableHeadTemplate/TableHeadTemplate";
-import {TableRowLanguages} from "@/components/view/TableRowLanguages/TableRowLanguages";
-import {TableTemplate} from "@/components/view/TableTemplate/TableTemplate";
+import {TableHeadTemplate} from "@view/TableHeadTemplate/TableHeadTemplate";
+import {TableRowLanguages} from "@/pages/Languages/TableRowLanguages/TableRowLanguages";
+import {TableTemplate} from "@view/TableTemplate/TableTemplate";
 import {Order} from "@/models/Order.type";
 import {LanguagesData} from "@/models/LanguagesData";
 import {LanguagesHeadCellsData} from "./LanguagesHeadCellsData";
