+++ src/pages/Positions/TablePositions/index.tsx
@@ -1,7 +1,7 @@
 import {FC} from "react";
 import {TableHeadTemplate} from "@view/TableHeadTemplate/TableHeadTemplate";
 import {TableTemplate} from "@view/TableTemplate/TableTemplate";
-import {TableRowPositions} from "@view/TableRowPositions";
+import {TableRowPositions} from "@/pages/Positions/TableRowPositions";
 import {PositionsData} from "@/models/PositionsData.type";
 import {Order} from "@/models/Order.type";
 import {useSort} from "@/hooks/useSort";
