+++ src/pages/Projects/Projects.tsx
@@ -1,5 +1,5 @@
 import {FC} from "react";
-import {TableProjects} from "@/components/containers/TableProjects/TableProjects";
+import {TableProjects} from "@/pages/Projects/TableProjects/TableProjects";
 import {useProjectsQuery} from "@/graphql/hooks/useProjectsQuery";
 
 export const Projects: FC = () => {
