+++ src/pages/Skills/TableSkills/skillsHeadCellsData.ts
